requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy[asyncio]>=2.0
asyncpg>=0.29
APScheduler>=3.10
pydantic[email]>=2.6
python-dotenv>=1.0
tzdata>=2024.1
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
packaging>=24.0
